refactor: delete data_connector_anonymizer
