+++ great_expectations/core/usage_statistics/anonymizers/datasource_anonymizer.py
@@ -1,9 +1,6 @@
 from typing import Optional
 
 from great_expectations.core.usage_statistics.anonymizers.anonymizer import Anonymizer
-from great_expectations.core.usage_statistics.anonymizers.data_connector_anonymizer import (
-    DataConnectorAnonymizer,
-)
 from great_expectations.datasource import (
     BaseDatasource,
     Datasource,
@@ -34,8 +31,6 @@ class DatasourceAnonymizer(Anonymizer):
             BaseDatasource,
         ]
 
-        self._data_connector_anonymizer = DataConnectorAnonymizer(salt=salt)
-
     def anonymize_datasource_info(self, name, config):
         anonymized_info_dict = {}
         anonymized_info_dict["anonymized_name"] = self.anonymize(name)
@@ -61,7 +56,7 @@ class DatasourceAnonymizer(Anonymizer):
             )
             data_connector_configs = config.get("data_connectors")
             anonymized_info_dict["anonymized_data_connectors"] = [
-                self._data_connector_anonymizer.anonymize_data_connector_info(
+                self.anonymize_data_connector_info(
                     name=data_connector_name, config=data_connector_config
                 )
                 for data_connector_name, data_connector_config in data_connector_configs.items()
@@ -106,7 +101,7 @@ class DatasourceAnonymizer(Anonymizer):
                         "module_name"
                     ] = "great_expectations.datasource.data_connector"
                 introspection_data_connector_anonymized_configs.append(
-                    self._data_connector_anonymizer.anonymize_data_connector_info(
+                    self.anonymize_data_connector_info(
                         name=data_connector_name, config=data_connector_config
                     )
                 )
@@ -126,7 +121,7 @@ class DatasourceAnonymizer(Anonymizer):
                         "module_name"
                     ] = "great_expectations.datasource.data_connector"
                 tables_data_connector_anonymized_configs.append(
-                    self._data_connector_anonymizer.anonymize_data_connector_info(
+                    self.anonymize_data_connector_info(
                         name=data_connector_name, config=data_connector_config
                     )
                 )
